fix: NSE cache_c, RECT relative imports, AlphaEdit float16 solve
Three vendor API bugs caught by GPU smoke tests:

1. NSE: needs cache_c initialized from model dims (not P). AlphaEdit
   and NSE have different requirements — AlphaEdit needs P+cache_c,
   NSE needs only cache_c.

2. RECT: baselines module uses relative imports (from .compute_ks).
   importlib.util doesn't set up parent package. Fix: temporarily
   swap sys.path to import from baselines memit package directly.

3. AlphaEdit via hooks: lhs was float16 when cov=None (skipped
   .double() conversion). torch.linalg.solve requires float64.
   Fix: explicit .double() before solve.

Added 4 CPU tests that catch these patterns.

+++ src/algorithms/alphaedit_with_hooks.py
@@ -122,8 +122,8 @@ def apply_alphaedit_with_hooks(
 
         # AlphaEdit solve: lhs @ upd = P @ K @ resid^T
         if hooks.build_lhs is not None:
-            upd_matrix = torch.linalg.solve(lhs, layer_ks)
-            upd_matrix = resid @ upd_matrix.T
+            upd_matrix = torch.linalg.solve(lhs.double(), layer_ks.double())
+            upd_matrix = resid.double() @ upd_matrix.T
         else:
             upd_matrix = torch.linalg.solve(lhs, P[i, :, :].cuda() @ layer_ks @ resid.T)
